Remove auction_models and registry_models packages
* Reduce model duplication
